+++ sdk/hdinsight/arm-hdinsight/src/hDInsightManagementClient.ts
@@ -75,7 +75,7 @@ export class HDInsightManagementClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-hdinsight/1.3.0-beta.1`;
+    const packageDetails = `azsdk-js-arm-hdinsight/2.0.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -128,7 +128,7 @@ export class HDInsightManagementClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-04-15-preview";
+    this.apiVersion = options.apiVersion || "2023-08-15-preview";
     this.applications = new ApplicationsImpl(this);
     this.clusters = new ClustersImpl(this);
     this.configurations = new ConfigurationsImpl(this);